spotify timestamp utc and stream default updates

+++ src/fitly/app.py
@@ -49,7 +49,7 @@ with server.app_context():
                 # scheduler.add_job(func=save_spotify_play_history, trigger="cron", minute='*/20')
 
                 # Use this job for polling every second (much more precise data with this method can detect skips, etc.)
-                scheduler.add_job(stream, "interval", seconds=int(config.get('spotify', 'poll_interval_seconds')))
+                scheduler.add_job(stream, "interval", seconds=float(config.get('spotify', 'poll_interval_seconds')))
 
             app.server.logger.info('Starting cron jobs')
             scheduler.start()
